update txn go ci test
