--- tests/integration/go-test/go.mod
@@ -1,16 +1,0 @@
-module go-test
-
-go 1.18
-
-require (
-	github.com/go-redis/redis/v8 v8.11.0
-	github.com/stretchr/testify v1.8.4
-)
-
-require (
-	github.com/cespare/xxhash/v2 v2.1.1 // indirect
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
-	github.com/pmezard/go-difflib v1.0.0 // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
-)
